refactor: split debug adapter types by concern
