+++ src/debug/index.ts
@@ -4,6 +4,9 @@
  */
 
 export * from './types';
+export * from './terminal-types';
+export * from './message-types';
+export * from './memory-types';
 export * from './breakpoint-manager';
 export * from './assembler';
 export * from './path-resolver';
