+++ src/platforms/tec1/provider.ts
@@ -14,7 +14,8 @@ import {
   handleSpeedRequest,
 } from '../../debug/platform-requests';
 import type { PlatformContribution } from '../../debug/platform-registry';
-import { extractKeyCode, type LaunchRequestArguments } from '../../debug/types';
+import type { LaunchRequestArguments } from '../../debug/types';
+import { extractKeyCode } from '../../debug/message-types';
 import type { PlatformCommandContext, ResolvedPlatformProvider } from '../provider';
 import { normalizeTec1Config } from './runtime';
